replace resulted indexes with actual names
