+++ .gitignore
@@ -2,4 +2,6 @@
 *.bak
 .firebase
 .firebase/*
+tmp
+
