+++ example/localtrust.csv
@@ -1,4 +1,4 @@
-0,1,11.31571
+alice,bob,11.31571
 2,3,269916.08616
 4,5,3173339.366896588
 6,5,46589750.00759474
