+++ example/pretrust.csv
@@ -1,1 +1,1 @@
-0,0.14285714285714285
+alice,0.14285714285714285
