+++ src/basic/engine.rs
@@ -9,11 +9,14 @@ use crate::basic::trustvector::canonicalize_trust_vector;
 use crate::basic::eigentrust::compute;
 use crate::basic::eigentrust::discount_trust_vector;
 
-pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Vec<Entry>, String> {
+pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Vec<(String, f64)>, String> {
     let e = 1.25e-7;
     let a = 0.5;
 
-    let (mut local_trust, mut peer_indices) = read_local_trust_from_csv(&localtrust_csv).unwrap();  
+    let (mut local_trust, mut peers) = read_local_trust_from_csv(&localtrust_csv).unwrap();  
+    
+    let mut peer_indices = peers.map;
+
     let mut pre_trust = read_trust_vector_from_csv(pretrust_csv, &peer_indices).unwrap();
 
     let c_dim = local_trust.cs_matrix.dim().unwrap();
@@ -23,27 +26,6 @@ pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Ve
     } else {
         pre_trust.set_dim(c_dim);
     }
-
-    /*
-    let lt_dim = local_trust.cs_matrix.entries.len();
-    let mut dim = lt_dim;
-
-    if peer_names.len() > 0 {
-        let n = peer_names.len();
-        if lt_dim < n {
-            dim = n;
-        } else if lt_dim > n {
-            return Err("localTrust is larger than peerNames".into());
-        }
-
-        let pt_dim = pre_trust.entries.len();
-        if pt_dim < n {
-            // Resize pre_trust if needed
-        } else if pt_dim > n {
-            return Err("preTrust is larger than peerNames".into());
-        }
-    }
-    */
     
     canonicalize_trust_vector(&mut pre_trust);
     
@@ -55,8 +37,10 @@ pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Ve
     let mut trust_scores = compute(&local_trust, &pre_trust, a, e, None, None).unwrap();
     discount_trust_vector(&mut trust_scores, &discounts)?;
 
-    /* 
+    
     let mut entries = vec![];
+
+    /*
     for i in 0..dim {
         let name = peer_names.as_ref().map_or_else(
             || format!("Peer {}", i),
@@ -64,8 +48,10 @@ pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Ve
         );
         entries.push(Entry::new(i, name));
     }
-
+    */
+    /*
     for e in &trust_scores {
+        let e = Entry{ value: e.score}
         entries[e.index].score = e.score;
         entries[e.index].score_log = e.score.log10();
     }
@@ -73,9 +59,10 @@ pub fn calculate_from_csv(localtrust_csv: &str, pretrust_csv: &str) -> Result<Ve
     entries.sort_by(|a, b| b.score.partial_cmp(&a.score).unwrap_or(std::cmp::Ordering::Equal));
     */
 
-    let mut entries = vec![];
     for e in &trust_scores.entries {
-        entries.push(Entry::new(e.index, e.value));
+        let name_ref = peers.map_reversed.get(&e.index).unwrap();
+        let name = name_ref.clone();
+        entries.push((name, e.value));
     }
 
     // entries.sort_by(|a, b| b.score.partial_cmp(&a.score).unwrap_or(std::cmp::Ordering::Equal));
