+++ src/basic/mod.rs
@@ -2,3 +2,4 @@ pub mod eigentrust;
 pub mod engine;
 pub mod localtrust;
 pub mod trustvector;
+pub mod util;
